Added base types because I think they look nicer.

+++ src/core.c
@@ -2,18 +2,18 @@
 
 #include <stdlib.h>
 
-void *malloc_stub(size_t size, void *ctx) {
+void *malloc_stub(u64 size, void *ctx) {
     (void) ctx;
     return malloc(size);
 }
 
-void free_stub(void *ptr, size_t size, void *ctx) {
+void free_stub(void *ptr, u64 size, void *ctx) {
     (void) ctx;
     (void) size;
     return free(ptr);
 }
 
-void *realloc_stub(void *ptr, size_t old, size_t new, void *ctx) {
+void *realloc_stub(void *ptr, u64 old, u64 new, void *ctx) {
     (void) ctx;
     (void) old;
     return realloc(ptr, new);
